102 Moves WfoSubscriptionProductBlock to its own folder
