+++ packages/orchestrator-ui-components/src/components/WfoSubscription/WfoSubscriptionProductBlock/index.ts
@@ -1,0 +1,1 @@
+export * from './WfoSubscriptionProductBlock';
